+++ src/client/java/spcore/imgui/nodes/types/math/Float/DivFloatMathNode.java
@@ -44,7 +44,7 @@ public class DivFloatMathNode extends AbstractNodeType {
             var2 = Float.parseFloat(node.getInputValue("var2"));
         }
 
-        outputs.put("value", (float)(var1 / var2));
+        outputs.put("value", var1 / var2);
         return outputs;
     }
 }
